requirements.txt
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
moviepy>=1.0.3
gtts>=2.3.2
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.97.0
